Move more elasticsearch types from data-types to the types package
