+++ packages/data-types/src/types/v1/boolean.ts
@@ -1,10 +1,15 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 export default class BooleanType extends BaseType {
     toESMapping(_version?: number) {
-        return { mapping: { [this.field]: { type: 'boolean' as ElasticSearchTypes } } };
+        return {
+            mapping: {
+                [this.field]: {
+                    type: 'boolean' as ESFieldType
+                }
+            }
+        };
     }
 
     toGraphQL() {
