+++ packages/data-types/src/types/v1/geo-json.ts
@@ -1,6 +1,5 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 export default class GeoJSON extends BaseType {
     toESMapping(_version?: number) {
@@ -9,7 +8,7 @@ export default class GeoJSON extends BaseType {
         return {
             mapping: {
                 [this.field]: {
-                    type: 'geo_shape' as ElasticSearchTypes,
+                    type: 'geo_shape' as ESFieldType,
                     tree: 'quadtree',
                     strategy: 'recursive'
                 }
