+++ packages/data-types/src/types/v1/geo.ts
@@ -1,11 +1,10 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType, { ToGraphQLOptions } from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 // TODO: This type is deprecated, not sure how to properly indicate it.
 export default class GeoType extends BaseType {
     toESMapping(_version?: number) {
-        return { mapping: { [this.field]: { type: 'geo_point' as ElasticSearchTypes } } };
+        return { mapping: { [this.field]: { type: 'geo_point' as ESFieldType } } };
     }
 
     toGraphQL({ isInput }: ToGraphQLOptions = {}) {
