+++ packages/data-types/src/types/v1/hostname.ts
@@ -1,17 +1,16 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 export default class Hostname extends BaseType {
     toESMapping(_version?: number) {
         return {
             mapping: {
                 [this.field]: {
-                    type: 'text' as ElasticSearchTypes,
+                    type: 'text' as ESFieldType,
                     analyzer: 'lowercase_keyword_analyzer',
                     fields: {
                         tokens: {
-                            type: 'text' as ElasticSearchTypes,
+                            type: 'text' as ESFieldType,
                             analyzer: 'hostname_analyzer',
                         },
                     },
