+++ packages/data-types/src/types/v1/ip.ts
@@ -1,10 +1,9 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 export default class IPType extends BaseType {
     toESMapping(_version?: number) {
-        return { mapping: { [this.field]: { type: 'ip' as ElasticSearchTypes } } };
+        return { mapping: { [this.field]: { type: 'ip' as ESFieldType } } };
     }
 
     toGraphQL() {
