+++ packages/data-types/src/types/v1/keyword-case-insensitive.ts
@@ -1,6 +1,5 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 export default class KeywordCaseInsensitive extends BaseType {
     toESMapping(_version?: number) {
@@ -15,7 +14,7 @@ export default class KeywordCaseInsensitive extends BaseType {
                         },
                     },
                 } : {
-                    type: 'text' as ElasticSearchTypes,
+                    type: 'text' as ESFieldType,
                     analyzer: 'lowercase_keyword_analyzer',
                 },
             },
