+++ packages/data-types/src/types/v1/ngram-tokens.ts
@@ -1,6 +1,5 @@
-import { xLuceneFieldType } from '@terascope/types';
+import { xLuceneFieldType, ESFieldType } from '@terascope/types';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../interfaces';
 
 export default class NgramTokens extends BaseType {
     toESMapping(_version?: number) {
@@ -8,10 +7,10 @@ export default class NgramTokens extends BaseType {
             mapping: {
                 [this.field]: {
                     // TODO: this is wrong, I dont think analyzer can be at this level
-                    type: 'keyword' as ElasticSearchTypes,
+                    type: 'keyword' as ESFieldType,
                     fields: {
                         tokens: {
-                            type: 'text' as ElasticSearchTypes,
+                            type: 'text' as ESFieldType,
                             analyzer: 'ngram_analyzer',
                         },
                     },
