+++ packages/data-types/test/types/v1/hostname-spec.ts
@@ -1,6 +1,7 @@
 import { TSError } from '@terascope/utils';
+import { ESFieldType } from '@terascope/types';
 import Hostname from '../../../src/types/v1/hostname';
-import { FieldTypeConfig, ElasticSearchTypes } from '../../../src/interfaces';
+import { FieldTypeConfig } from '../../../src/interfaces';
 
 describe('Hostname V1', () => {
     const field = 'someField';
@@ -28,11 +29,11 @@ describe('Hostname V1', () => {
         const results = {
             mapping: {
                 [field]: {
-                    type: 'text' as ElasticSearchTypes,
+                    type: 'text' as ESFieldType,
                     analyzer: 'lowercase_keyword_analyzer',
                     fields: {
                         tokens: {
-                            type: 'text' as ElasticSearchTypes,
+                            type: 'text' as ESFieldType,
                             analyzer: 'hostname_analyzer',
                         },
                     },
